Refine RecipeHeader metadata flow
